chore: Renamed service/kernel/kernel.service.mts to service/kernel/execution-context.registry.mts.
