+++ packages/architectura/src/service/kernel/_index.mts
@@ -1,1 +1,1 @@
-export * from './kernel.service.mjs';
+export * from './execution-context.registry.mjs';
