+++ packages/architectura/src/service/kernel/execution-context.registry.mts
